requirements.txt
discord.py==2.3.2
fastapi==0.110.2
uvicorn[standard]==0.29.0
httpx==0.27.0
pyyaml==6.0.1
pydantic==2.6.4
python-dotenv==1.0.1
jinja2==3.1.3
